+++ pilz_trajectory_generation/CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog for package pilz_trajectory_generation
 
 Forthcoming
 -----------
+* add append method for avoiding duplicate points in robot_trajectory trajectories
+* Relax the precondition on trajectory generators from v_start==0 to |v_start| < 1e-10 to gain robustness
 * Set last point of generated trajectories to have vel=acc=0 to match the first point.
 * add sequence action and service capabilities to concatenate multiple requests
 * Contributors: Pilz GmbH and Co. KG